moved to the utils/material folder | 'cosmetics' | TODO: Lazy<Ingredient> should be updated as soon as possible!
